# OTT-Platform-Web-Application/Backend/docs/api_docs.yaml
...
openapi: 3.0.3
info:
  title: Dhaakad API
  description: Dhaakad API documentation (Might be buggy)
  version: 0.0.5
  contact:
    email: vishal.rustagi@metaorangedigital.com
servers:
  - url: 'https://api.dhaakadcinema.com/api/v1'
...
paths:
...
  /reviews/updatestatus:
    post:
      tags:
        - reviews
      summary: Update Status
      operationId: updateStatus
      parameters:
        - name: reviewId
          in: query
          required: true
          schema:
            type: string
            example: xyz123
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                  example: 'abc123'
                review:
                  type: string
                  example: 'Good Movie'
                description:
                  type: string
                  example: 'Good Movie'
                rating:
                  type: integer
                  example: 1||2||3||4||5
                movieId:
                  type: string
                  example: 'asdasd'
                isApproved:
                  type: boolean
                  example: true
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "updated successfully"
                  data:
                    type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Review'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: [ ]
  /categories/:
    get:
      tags:
        - categories
      summary: get categories
      operationId: getCategories
      responses:
        201:
          description: category fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "All categories fetch successfully"
                  result:
                    type: number
                    example: 3
                  data:
                    type: object
                    properties:
                      category:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - categories
      summary: Create new Category
      operationId: createCategory
      requestBody:
        description: create cast object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        201:
          description: category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "category created successfully"
                  data:
                    type: object
                    properties:
                      category:
                        $ref: '#/components/schemas/Category'
      security:
        - bearerAuth: [ ]
  /categories/{id}:
    get:
      tags:
        - categories
      summary: get category by id
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          description: id of the cateogry
          required: true
          schema:
            type: string
            example: 'aajsndjkansd'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "get category successfully"
                  data:
                    type: object
                    properties:
                      category:
                        $ref: '#/components/schemas/Category'
      security:
        - bearerAuth: [ ]
    patch:
      tags:
        - categories
      summary: update category
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          description: id of the category
          required: true
          schema:
            type: string
            example: 'aajsndjkansd'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "category updated successfully"
                  data:
                    type: object
                    properties:
                      category:
                        $ref: '#/components/schemas/Category'
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - categories
      summary: delete category
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          description: id of the category
          required: true
          schema:
            type: string
            example: 'aajsndjkansd'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "category deleted successfully"
                  data:
                    type: object
                    properties:
                      cast:
                        $ref: '#/components/schemas/Category'
        404:
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'error'
                  message:
                    type: string
                    example: 'No category found with that ID'

      security:
        - bearerAuth: [ ]
  /languages:
    get:
      tags:
        - languages
      summary: get all languages
      operationId: getAllLanguages
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "All languages fetch successfully"
                  results:
                    type: integer
                    example: 10
                  data:
                    type: object
                    properties:
                      language:
                        type: array
                        items:
                          $ref: '#/components/schemas/Language'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - languages
      summary: Add language
      operationId: addLanguage
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 'English'
                status:
                  type: string
                  example: 'Active||Inactive'
      responses:
        201:
          description: Language created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Language created successfully"
                  data:
                    type: object
                    properties:
                      language:
                        $ref: '#/components/schemas/Language'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Review:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'xyz123'
        review:
          type: string
          example: 'Good Movie'
        description:
          type: string
          example: 'Good Movie'
        rating:
          type: integer
          example: 1||2||3||4||5
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        movieId:
          type: string
          example: 'asdasd'
        user:
          type: string
          example: 'abc123'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        isApproved:
          type: boolean
          example: false
    Category:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'asdasd'
        title:
          type: string
          uniqueItems: true
          example: 'Action'
        status:
          type: string
          example: 'Active'
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
    Language:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'asd456'
        title:
          type: string
          example: English
        status:
          type: string
          example: 'Active||Inactive'
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: 'Something went wrong'
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes